+++ armada_command/docker_utils/compatibility.py
@@ -13,7 +13,7 @@ DOCKER_STATIC_CLIENT_DIR = '/opt/armada-docker-client/'
 
 
 def get_docker_version():
-    output = check_output(['docker', '--version'])
+    output = check_output(['docker', '--version']).decode("utf-8")
     match = re.search(r'^Docker version (?P<version>\d+\.\d+\.\d+)', output)
     try:
         return match.group('version')